pyproject.toml
[build-system]
requires = ["setuptools >= 64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "map_utils"
version = "0.1"
dependencies = [
    "numpy==2.0.2",
    "mygene==3.2.2",
    "scipy==1.15.1",
    "pandas==2.2.3",
    "copairs==0.5.1",
    "notebook==7.3.2",
    "seaborn==0.13.2",
    "networkx==3.4.2",
    "plotnine==0.14.5",
    "hydra-core==1.3.2",
    "snakemake==8.27.1",
    "umap-learn==0.5.7",
    "pycytominer==1.2.2",
    "scikit-learn==1.6.1",
    "statsmodels==0.14.4",
]
requires-python = "==3.11"
authors = [
  {name = "Alexandr Kalinin", email = "akalinin@broadinstitute.org"},
]
description = "Utilities for the (Kalinin et al., 2024) mAP paper"
readme = "README.md"
license = {file = "LICENSE"}

[tool.setuptools.packages.find]
where = ["."]
include = ["map_utils*"]

[tool.setuptools.package-data]
map_utils = ["fonts/*.ttf"]

[project.optional-dependencies]
dev = [
    "ruff"
]
